Crate stivale

A crate for parsing qloader2 and tomatboot’s stivale2 structures

The header module contains a header struct and tags for letting a bootloader know that a kernel is stivale2 compliant

StivaleStructure is loaded with load, with an address that’s passed in RDI on x86_64 (the first function parameter on an extern "C" fn)

Re-exports

pub use header::*;

Modules

epoch

firmware

framebuffer

header

stivale2 header module for letting a stivale2 compliant bootloader find the kernel

memory

module

rsdp

terminal

Structs

StivaleStructure

The stivale2 structure containing all the tags passed by the bootloader

StivaleStructureInner

Functions

load⚠

Load the stivale2 structure from an address