Struct stivale::StivaleStructure

pub struct StivaleStructure { /* fields omitted */ }

The stivale2 structure containing all the tags passed by the bootloader

Implementations

impl StivaleStructure

pub fn bootloader_brand(&self) -> Option<&str>

Get the bootloader brand that booted the kernel, if any

pub fn bootloader_version(&self) -> Option<&str>

Get the bootloader version, if any

pub fn framebuffer(&self) -> Option<&'static FramebufferTag>

Get the video framebuffer info tag

pub fn rsdp(&self) -> Option<&'static RSDPTag>

Get the ACPI RSDP structure pointer

pub fn epoch(&self) -> Option<&'static EpochTag>

Get the current UNIX epoch during boot

pub fn firmware(&self) -> Option<&'static FirmwareTag>

Get the firmware tag passed by the bootloader

pub fn memory_map(&self) -> Option<&'static MemoryMapTag>

Get the memory map tag

pub fn terminal(&self) -> Option<&'static TerminalTag>

Get the terminal tag

pub fn module(&self) -> Option<&'static ModuleTag>

Get the module tag